release: prepare Topbase 0.2.4

+++ internal/core/dashboard.go
@@ -3,10 +3,10 @@ package core
 import "time"
 
 type Layout struct {
-	X int `json:"x"`
-	Y int `json:"y"`
-	W int `json:"w"`
-	H int `json:"h"`
+	X float64 `json:"x"`
+	Y float64 `json:"y"`
+	W float64 `json:"w"`
+	H float64 `json:"h"`
 }
 
 type ClickBehavior struct {
